# perf: reduce the memory usage of LC Trie construction