--- a/source/common/network/lc_trie.h
+++ b/source/common/network/lc_trie.h
@@ -452,6 +452,8 @@ private:
 
       // The value of next_free_index is the final size of the trie_.
       ASSERT(next_free_index <= trie_.size());
+      trie_.resize(next_free_index);
+      trie_.shrink_to_fit();
     }
 
     // Thin wrapper around computeBranch output to facilitate code readability.