--- a/test/common/network/lc_trie_speed_test.cc
+++ b/test/common/network/lc_trie_speed_test.cc
@@ -12,9 +12,14 @@ std::vector<std::pair<std::string, std::vector<Envoy::Network::Address::CidrRang
 std::vector<std::pair<std::string, std::vector<Envoy::Network::Address::CidrRange>>>
     tag_data_nested_prefixes;
 
+std::vector<std::pair<std::string, std::vector<Envoy::Network::Address::CidrRange>>>
+    tag_data_minimal;
+
 std::unique_ptr<Envoy::Network::LcTrie::LcTrie<std::string>> lc_trie;
 
 std::unique_ptr<Envoy::Network::LcTrie::LcTrie<std::string>> lc_trie_nested_prefixes;
+
+std::unique_ptr<Envoy::Network::LcTrie::LcTrie<std::string>> lc_trie_minimal;
 
 } // namespace
 
@@ -41,12 +46,10 @@ static void BM_LcTrieConstructNested(benchmark::State& state) {
 BENCHMARK(BM_LcTrieConstructNested);
 
 static void BM_LcTrieConstructMinimal(benchmark::State& state) {
-  std::vector<std::pair<std::string, std::vector<Envoy::Network::Address::CidrRange>>> tags;
-  tags.emplace_back(std::pair<std::string, std::vector<Envoy::Network::Address::CidrRange>>(
-      {"tag_1", {Envoy::Network::Address::CidrRange::create("0.0.0.0/0")}}));
+
   std::unique_ptr<Envoy::Network::LcTrie::LcTrie<std::string>> trie;
   for (auto _ : state) {
-    trie = std::make_unique<Envoy::Network::LcTrie::LcTrie<std::string>>(tags);
+    trie = std::make_unique<Envoy::Network::LcTrie::LcTrie<std::string>>(tag_data_minimal);
   }
   benchmark::DoNotOptimize(trie);
 }
@@ -79,6 +82,19 @@ static void BM_LcTrieLookupWithNestedPrefixes(benchmark::State& state) {
 
 BENCHMARK(BM_LcTrieLookupWithNestedPrefixes);
 
+static void BM_LcTrieLookupMinimal(benchmark::State& state) {
+  static size_t i = 0;
+  size_t output_tags = 0;
+  for (auto _ : state) {
+    i++;
+    i %= addresses.size();
+    output_tags += lc_trie_minimal->getData(addresses[i]).size();
+  }
+  benchmark::DoNotOptimize(output_tags);
+}
+
+BENCHMARK(BM_LcTrieLookupMinimal);
+
 } // namespace Envoy
 
 // Boilerplate main(), which discovers benchmarks in the same file and runs them.
@@ -92,9 +108,10 @@ int main(int argc, char** argv) {
     addresses.push_back(Envoy::Network::Utility::parseInternetAddress(address));
   }
 
-  // Construct two sets of prefixes: one consisting of 1,024 addresses in an
-  // RFC 5737 netblock, the other consisting of those same addresses plus
-  // 0.0.0.0/0 (to exercise the LC Trie's support for nested prefixes)
+  // Construct three sets of prefixes: one consisting of 1,024 addresses in an
+  // RFC 5737 netblock, another consisting of those same addresses plus
+  // 0.0.0.0/0 (to exercise the LC Trie's support for nested prefixes),
+  // and finally a set containing only 0.0.0.0/0.
   for (int i = 0; i < 32; i++) {
     for (int j = 0; j < 32; j++) {
       tag_data.emplace_back(std::pair<std::string, std::vector<Envoy::Network::Address::CidrRange>>(
@@ -106,10 +123,14 @@ int main(int argc, char** argv) {
   tag_data_nested_prefixes.emplace_back(
       std::pair<std::string, std::vector<Envoy::Network::Address::CidrRange>>(
           {"tag_0", {Envoy::Network::Address::CidrRange::create("0.0.0.0/0")}}));
+  tag_data_minimal.emplace_back(
+      std::pair<std::string, std::vector<Envoy::Network::Address::CidrRange>>(
+          {"tag_1", {Envoy::Network::Address::CidrRange::create("0.0.0.0/0")}}));
 
   lc_trie = std::make_unique<Envoy::Network::LcTrie::LcTrie<std::string>>(tag_data);
   lc_trie_nested_prefixes =
       std::make_unique<Envoy::Network::LcTrie::LcTrie<std::string>>(tag_data_nested_prefixes);
+  lc_trie_minimal = std::make_unique<Envoy::Network::LcTrie::LcTrie<std::string>>(tag_data_minimal);
 
   benchmark::Initialize(&argc, argv);
   if (benchmark::ReportUnrecognizedArguments(argc, argv)) {