*Description*:
Rather than preallocating a very large buffer to hold the LC Trie nodes
the way the reference implementation from the original  Nilsson & Karlsson
paper did, the trie construction now starts out with a size based on the
number of prefixes and lets the `std::vector` manage any needed reallocation
during trie construction.

In addition to using less memory during trie construction, this new approach
ended up being faster:

```
Benchmark                                  Time           CPU Iterations
-------------------------------------------------------------------------
Before:
BM_LcTrieConstruct                  26631971 ns   26522462 ns         26

After:
BM_LcTrieConstruct                   3654633 ns    3646942 ns        189
```

*Risk Level*: medium
*Testing*: unit tests plus new speed benchmark test
*Docs Changes*: N/A
*Release Notes*: N/A

Signed-off-by: Brian Pane <bpane@pinterest.com>