@brian-pane 2018-08-12
@PiotrSikora @rshriram @ccaraman this PR is a follow-up to [3980](https://github.com/envoyproxy/envoy/pull/3890) that aims to further reduce the temporary memory footprint during trie construction.